+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/search/user/SearchUsers.kt
@@ -49,6 +49,7 @@ fun SearchUsers(
             state = listState,
         ) {
             item(key = "aboveSearch") { aboveSearch() }
+            stickyHeader("searchErrors") { SearchErrors(userSearchViewModel) }
             stickyHeader("searchTerm") {
                 val containerColor = LocalSurfaceStyle.current?.color
                 Column((if (containerColor != null) Modifier.background(containerColor) else Modifier)) {